Add support for a "smart" Stop Button.
Most modern players don't really have a concept of Stop with the exception of players that play continuous streams that can't be paused (like internet radio). The Stop Button will only be shown if CanPause is false and canPlay is true. In the event that the Stop is Shown the PlayPause Button turns to just a PlayButton.