Add support for shuffle and repeat and a few other unrelated tweaks and fixes
Add a shuffle and repeat button to the UI for players that the Shuffle or Loopstatus props work correctly (and even exist, they are optional).

For the sake of UI symmetry if either Shuffle or Loopstatus works both buttons will be shown, the button shown for whichever non-functional prop will just be non-reactive.