HUGE Updates
The update will include:

initial 3.34 support
TrackList interface support
PlayLists interface support
A quit button
Better widget alignment
A separator between players
An arrow to indicate the current player if there are more than one in the menu
Focusing a player will skip the annoying focus stealing prevention "Player is Ready" popup notification and just focus the damn player.

Plus a bunch of internal changes.

This update is not considered stable and ready for general release but is good enough for the brave to start testing.