Fix for theme that set the size of icons in the aggregate-menu style class in height and width.
It is possible for a theme to set the size of the cover by specifying height and width along with or instead of icon-size. Which can result in the icon not being the correct size or out of proportion. This prevents any style classes from mangling our cover icon's size.